fix(item-info): code moved to parser to allow further usage in other UI
