+++ renderer/src/parser/ParsedItem.ts
@@ -66,6 +66,11 @@ export interface ParsedItem {
   }
   category?: ItemCategory
   info: BaseType
+  disenchantCandidates: Array<{
+    name: string
+    value: string
+    icon: string
+  }>
   rawText: string
 }
 
@@ -86,6 +91,7 @@ export function createVirtualItem (
     statsByType: props.statsByType ?? [],
     unknownModifiers: props.unknownModifiers ?? [],
     influences: props.influences ?? [],
+    disenchantCandidates: props.disenchantCandidates ?? [],
     rawText: 'VIRTUAL_ITEM'
   }
 }
